feat(resources): extract list components for materials and indents (Phase 6a)
- Create 'features/resources/components' barrel for presentational components.
- Extract 'MaterialListItem' and 'IndentRow' into dedicated components.
- Refactor Materials and Indents pages to use these extracted components.
- Reduce page complexity by delegating row/item rendering.
